requirements.txt
Django==2.2.6
django-dynamic-fixtpip ure==2.0.0
djangorestframework==3.10.3
